05-security-agent/requirements.txt
anthropic>=0.28.0
python-dotenv>=1.0.0
requests
OTXv2